small changes to types, adding functionality GetIngredients to the Dish file

+++ OrderingSystem/Ingredient.cs
@@ -9,9 +9,9 @@ namespace OrderingSystem
     public class Ingredient
     {   
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public Ingredient(string name, double price)
+        public Ingredient(string name, decimal price)
         {
             Name = name;
             Price = price;
